+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added version filter on `Get-Instance` command.
+
 ## [0.3.0] - 2022-09-30
 
 ### Added