+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2023-03-07
+
 ### Added
 
 - Added version filter on `Get-Instance` command.
+
+### Fixed
+
+- Empty version parameter used as none version.
 
 ## [0.3.0] - 2022-09-30